refactor: move pure openapi-spec-enrichment helper to infra/utils
constructs/ is reserved for CDK construct/stack classes; this module is CDK-free.
