+++ packages/back-end/src/infra/utils/openapi-spec-enrichment.ts
